requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
clip==0.2.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
lightning==2.5.1
lightning-utilities==0.14.3
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.3.2
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
propcache==0.3.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchmetrics==1.7.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.13.1
urllib3==2.3.0
yarl==1.18.3
